Rename file extension from .js to .jsx
